third_party/WebKit/LayoutTests/http/tests/devtools/console/console-viewport-indices-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verifies viewport's visible and active message ranges.


Running: testEmptyViewport
Expected and actual visible ranges match

Running: testFirstLastVisibleIndices
Logging 100 Normal messages
Force item to be first: 0
Expected and actual visible ranges match
Force item to be first: 1
Expected and actual visible ranges match
Scroll a bit down: 15px
Expected and actual visible ranges match
Force item to be last: 50
Expected and actual visible ranges match
Force item to be last: 99
Expected and actual visible ranges match

Running: testMultilineMessages
Logging 100 Multiline messages
Force item to be first: 0
Expected and actual visible ranges match
Force item to be first: 1
Expected and actual visible ranges match
Scroll a bit down: 15px
Expected and actual visible ranges match
Force item to be last: 50
Expected and actual visible ranges match
Force item to be last: 99
Expected and actual visible ranges match

Running: testSlightlyBiggerMessages
Logging 100 SlightlyBigger messages
Force item to be first: 0
Expected and actual visible ranges match
Force item to be first: 1
Expected and actual visible ranges match
Scroll a bit down: 15px
Expected and actual visible ranges match
Force item to be last: 50
Expected and actual visible ranges match
Force item to be last: 99
Expected and actual visible ranges match
